# commands/cluster: use pipeline to execute split commands