--- a/redis/commands/__init__.py
+++ b/redis/commands/__init__.py
@@ -1,4 +1,4 @@
-from .cluster import AsyncRedisClusterCommands, RedisClusterCommands
+from .cluster import READ_COMMANDS, AsyncRedisClusterCommands, RedisClusterCommands
 from .core import AsyncCoreCommands, CoreCommands
 from .helpers import list_or_args
 from .parser import CommandsParser
@@ -6,14 +6,15 @@ from .redismodules import AsyncRedisModuleCommands, RedisModuleCommands
 from .sentinel import AsyncSentinelCommands, SentinelCommands
 
 __all__ = [
-    "AsyncRedisClusterCommands",
-    "RedisClusterCommands",
-    "CommandsParser",
     "AsyncCoreCommands",
-    "CoreCommands",
-    "list_or_args",
+    "AsyncRedisClusterCommands",
     "AsyncRedisModuleCommands",
-    "RedisModuleCommands",
     "AsyncSentinelCommands",
+    "CommandsParser",
+    "CoreCommands",
+    "READ_COMMANDS",
+    "RedisClusterCommands",
+    "RedisModuleCommands",
     "SentinelCommands",
+    "list_or_args",
 ]